src/combat_agent_refactored/pyproject.toml
[project]
name = "dnd-combat-simulator"
version = "1.0.0"
description = "Turn-based DnD combat simulator with AI-powered enemies"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "DnD Master Team" }
]

dependencies = [
    "google-genai>=1.39.1",
    "openai>=1.95.1",
    "numpy>=1.24.0",
    "pandas>=2.3.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
dnd-combat = "cli:main"
dnd-db = "db_tool:main"

[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["combat_engine", "db_tool", "cli"]
